% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markDelivery.R
\name{deliveryThreshold}
\alias{deliveryThreshold}
\title{Delivery Count Threshold}
\usage{
deliveryThreshold(data, daylist, cts = getOption("pa.cts"), dist = c("t",
  "normal"), stat = c("mean", "sd", "95"))
}
\description{
Delivery Count Threshold
}
\keyword{internal}

