% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markDelivery.R
\name{markDelivery}
\alias{markDelivery}
\title{Mark Delivery Days}
\usage{
markDelivery(data, cts = getOption("pa.cts"), markingString = "w",
  window = c("trim", "consecutive", "valid"), method = c("95", "mean",
  "sd"), validCut = getOption("pa.validCut"), wearThreshold = 300,
  dist = c("t", "normal"))
}
\arguments{
\item{data}{Data with classified wear (nonwear) status by
\code{\link{wearingMarking}}.}

\item{cts}{The name of the counts column. The default is "axis1".}

\item{markingString}{Option for summarizing wear (markingString = "w") or
nonwear time (markingString = "nw").}

\item{window}{character.  Should be one of 'trim', 'consecutive', or 'valid'.}

\item{method}{character.  Should be one of '95', 'mean' or 'sd'.}

\item{validCut}{A cutoff for the total minutes of classified monitor wear
time per day to be considered as a valid monitor day.}

\item{wearThreshold}{A numeric value specifying a pseudo-valid day cutoff
similar to "validCut".}

\item{dist}{A distribution used to calculate the 95\% CI.}
}
\value{
A dataframe with summary information about daily counts.
}
\description{
This function adds an indicator variable for accelerometer delivery days based
on a delivery classification algorithm. The algorithm classifies each day as 
delivery or non-delivery day within each participant data using summary 
statistics of accelerometer counts for each day. As the summary statistics,
the 95th percentile, mean and standard deviation (sd) of accelerometer counts
can be used. With the summary statistics, the algorithm defines a set of days 
using 3 methods (namely \code{trim}, \code{consecutive}, and \code{valid}),
which are used to estimate the 95\% confidence interval (CI) based on
$t$-distribution (default) or normal distribution. The lower bound of
the 95\%  CI is used to classify delivery days; that is if the summary 
statistics for a day is below the lower bound of the 95\% CI, this day is 
classified as delivery day.
}
\examples{
data(deliveryData)

options(pa.cts = "vm")
wm <- wearingMarking(dataset = deliveryData)

markDelivery(wm)
plotData(data=wm) # days 1, 2, 10 - 15 are delivery or invalid days based on the result above
markDelivery(wm, window='valid', method='mean')
markDelivery(wm, method='mean')
markDelivery(wm, method='sd')
}
