% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markPAI.R
\name{markPAI}
\alias{markPAI}
\title{Mark Physical Activity Intensity (PAI) Level}
\usage{
markPAI(data, cts = getOption("pa.cts"), markingString = "w",
  breaks = c(-Inf, 100, 760, 2020, Inf), labels = c("sedentary", "light",
  "moderate", "vigorous"))
}
\arguments{
\item{data}{Data with classified wear (nonwear) status by
\code{\link{wearingMarking}}.}

\item{cts}{The name of the counts column. The default is "axis1".}

\item{markingString}{Option for summarizing wear (markingString = "w") or
nonwear time (markingString = "nw").}

\item{breaks}{A numeric vector of cut-points.}

\item{labels}{A character vector labelling intensity levels.}
}
\value{
A dataframe with an additional pai-level column.
}
\description{
This function adds a physical activity intensity level variable to the
source dataset.  "pai" is an ordered factor variable.  It will be NA for
nonwear times.
}
\examples{
data(dataSec)

mydata1m = dataCollapser(dataSec, TS = "TimeStamp", col = "counts", by = 60)

data1m = wearingMarking(dataset = mydata1m,
                       perMinuteCts = 1,
                       cts = "counts")

markPAI(data = data1m, cts = 'counts')[1:10,]
}

