% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readActigraph.R
\name{modeNames}
\alias{modeNames}
\title{Retrieve ActiGraph Column Names}
\usage{
modeNames(mode)
}
\description{
Column names should match the mode number.
}
\details{
https://actigraph.desk.com/customer/en/portal/articles/
2515800-what-do-the-different-mode-numbers-mean-in-a-csv-or-dat-file-
}
\keyword{internal}

