\name{nthOccurance}
\alias{nthOccurance}
\title{Internal Function}
\description{
This is an internal function, not for users.
}
\usage{
nthOccurance(dataVct, value, nth = NA, reverse = FALSE)
}


\arguments{
  \item{dataVct}{dataVct}
  \item{value}{value}
  \item{nth}{nth}
  \item{reverse}{reverse}
}

\value{
loc
}

\author{Zhouwen Liu <zhouwen.liu@Vanderbilt.Edu>}

\keyword{internal}
