% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pa-internal}
\alias{pa-internal}
\alias{l2df}
\alias{convertTZ}
\alias{createTime}
\alias{addDayIndex}
\title{Internal functions}
\description{
code{l2df}: convert list to data.frame
}
\details{
code{convertTZ}: force a new timezone on datetime

code{createTime}: create datetime variable

code{addDayIndex}: create ID_day variable from timestamp
}
\keyword{internal}
