% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumVct.R
\name{sumVct}
\alias{sumVct}
\title{Summarize Wearing or Nonwearing Time Interval}
\usage{
sumVct(datavct, wearing = "wearing", TS = getOption("pa.timeStamp"),
  markingString = "w", by = "days", id = NULL)
}
\arguments{
\item{datavct}{Data with classified wear (nonwear) status classified by
\code{\link{wearingMarking}}.}

\item{wearing}{The column name for classified wear and nonwear status. The
default is "wearing".}

\item{TS}{The column name for timestamp. The default is "TimeStamp".}

\item{markingString}{Option for summarizing wear (markingString="w") or
nonwear time interval (markingString="nw").}

\item{by}{A sequence of days for classified wear (nonwear) time intervals.}

\item{id}{Optional output for subject identification or file name.}
}
\value{
The summary data for wearing or nonwearing time intervals.
}
\description{
This function summarizes the classified wear (nonwear) time by interval basis
from the epoch-by-epoch classified wear (nonwear) status classified by
\code{\link{wearingMarking}}.
}
\examples{

data(dataSec)

mydata1m = dataCollapser(dataSec, TS = "TimeStamp", col = "counts", by = 60)

data1m = wearingMarking(dataset = mydata1m,
                       frame = 90, 
                       perMinuteCts = 1,
                       TS = "TimeStamp",
                       cts = "counts", 
                       streamFrame = NULL, 
                       allowanceFrame= 2, 
                       newcolname = "wearing")

sumVct(data1m, id="sdata1m")
sumVct(data1m, id="sdata1m", markingString = "nw")
}
\author{
Leena Choi \email{leena.choi@Vanderbilt.Edu},
Zhouwen Liu \email{zhouwen.liu@Vanderbilt.Edu},
Charles E. Matthews \email{Charles.Matthews2@nih.gov}, and
Maciej S. Buchowski \email{maciej.buchowski@Vanderbilt.Edu}
}
\references{
Choi, L., Liu, Z., Matthews, C. E. and Buchowski, M. S.  (2010).
Validation of Accelerometer Wear and Nonwear Time Classification Algorithm.
Med Sci Sports Exerc. Jun 23 [Epub ahead of print].
}
\seealso{
\code{\link{wearingMarking}}, \code{\link{summaryData}}
}

