% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readActigraph.R
\name{timeFromYear1}
\alias{timeFromYear1}
\title{Create Time From Seconds Since 0001-01-01}
\usage{
timeFromYear1(x, base = 62135596800, tz = "UTC")
}
\description{
There are 62,135,596,800 seconds from 0001-01-01 until 1970-01-01.
The last 1e7 digits of x are unnecessary microseconds.
}
\keyword{internal}
