% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyCounts.R
\name{verifyCounts}
\alias{verifyCounts}
\title{Verify Daily Counts}
\usage{
verifyCounts(data, summary, verifyFUN = mean, cts = "axis1",
  markingString = "w", nonzero = FALSE, conf = 0.95)
}
\arguments{
\item{data}{Data with classified wear (nonwear) status by
\code{\link{wearingMarking}}.}

\item{summary}{List output of \code{\link{summaryData}} function.}

\item{verifyFUN}{Function to describe daily counts.  The default is
\code{mean}.}

\item{cts}{The name of the counts column. The default is "axis1".}

\item{markingString}{Option for summarizing wear (markingString = "w") or
nonwear time (markingString = "nw").}

\item{nonzero}{Option to exclude zero value counts from the summary.}

\item{conf}{Confidence interval, defaulting to 95%.  Use NA to return
probabilites.}
}
\value{
Logical vector or vector of probabilites if conf is NA.
}
\description{
Some days may be incorrectly classified as valid.  This function attempts
to verify those dates.
}
\examples{

data(dataSec)

mydata1m = dataCollapser(dataSec, TS = "TimeStamp", col = "counts", by = 60)

data1m = wearingMarking(dataset = mydata1m,
                       frame = 90, 
                       perMinuteCts = 1,
                       TS = "TimeStamp",
                       cts = "counts", 
                       streamFrame = NULL, 
                       allowanceFrame= 2, 
                       newcolname = "wearing")
sumdat <- summaryData(data=data1m)
verifyCounts(data1m, sumdat, verifyFUN = mean, cts = "counts")
verifyCounts(data1m, sumdat, verifyFUN = median, cts = "counts",
             nonzero = TRUE, conf = NA)
}
\seealso{
\code{\link{wearingMarking}}, \code{\link{summaryData}}
}

