% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSec.R
\docType{data}
\name{dataSec}
\alias{dataSec}
\title{Accelerometer Data Example}
\format{A data frame with 238140 observations on the following 2 required
variables.
\describe{
  \item{TimeStamp}{a character vector, timestamp of accelerometer measurements}
  \item{counts}{a numeric vector, counts as accelerometer measurements}
}}
\usage{
dataSec
}
\description{
Approximately 3 days of accelerometer data collected with 1-sec epoch in the
correct data format that can be used by \code{wearingMarking} to classify
wear and nonwear time.
}
\examples{
data(dataSec)
}
\references{
Choi L, Liu Z, Matthews CE, Buchowski MS. 
Validation of accelerometer wear and nonwear time classification algorithm.
Med Sci Sports Exerc. 2011 Feb;43(2):357-64.
}
\keyword{datasets}
