% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSec.R
\docType{data}
\name{dataSec}
\alias{dataSec}
\title{Accelerometer Data Example}
\format{A data frame with 238140 observations on the following 2 required variables.
\describe{
  \item{TimeStamp}{a character vector}
  \item{counts}{a numeric vector}
}}
\usage{
dataSec
}
\description{
Approximately 3 days of accelerometer data collected with 1-sec epoch in the
correct data format that can be used by \code{wearingMarking} to classify
wear and nonwear time.
}
\examples{
data(dataSec)
}
\references{
Choi, L., Liu, Z., Matthews, C. E. and Buchowski, M. S.  (2010).
Validation of Accelerometer Wear and Nonwear Time Classification Algorithm.
Med Sci Sports Exerc. Jun 23 [Epub ahead of print].
}
\keyword{datasets}

