% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliveryData.R
\docType{data}
\name{deliveryData}
\alias{deliveryData}
\title{Delivery Classification Data Example}
\format{A data frame with 20987 observations on the following 6 required
variables.
\describe{
  \item{TimeStamp}{a character vector}
  \item{axis1}{a numeric vector}
  \item{axis2}{a numeric vector}
  \item{axis3}{a numeric vector}
  \item{steps}{a numeric vector}
  \item{vm}{a numeric vector, vector magnitude}
}}
\usage{
deliveryData
}
\description{
Approximately 15 days of accelerometer data collected with 1-minute epoch.
}
\examples{
data(deliveryData)
}
\keyword{datasets}

