% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliveryFeatures.R
\name{deliveryFeatures}
\alias{deliveryFeatures}
\title{deliveryFeatures}
\usage{
deliveryFeatures(df, ...)
}
\arguments{
\item{df}{details}

\item{\dots}{possibly used to create datetime variable from TimeStamp}
}
\value{
details
}
\description{
description
}
\details{
details
}
