% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliveryPred.R
\name{deliveryPred}
\alias{deliveryPred}
\title{deliveryPred}
\usage{
deliveryPred(df, model = c("RF", "NN", "GLM"))
}
\arguments{
\item{df}{details}

\item{model}{details}
}
\value{
details
}
\description{
description
}
\details{
details
}
