% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliveryPrediction.R
\name{deliveryPrediction}
\alias{deliveryPrediction}
\title{deliveryPrediction}
\usage{
deliveryPrediction(df, feats, model = c("RF", "GLM", "NN"), ...)
}
\arguments{
\item{df}{details}

\item{feats}{details}

\item{model}{details}

\item{\dots}{possibly used to create datetime variable from TimeStamp}
}
\value{
details
}
\description{
description
}
\details{
details
}
