% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliveryPreprocess.R
\name{deliveryPreprocess}
\alias{deliveryPreprocess}
\title{deliveryPreprocess}
\usage{
deliveryPreprocess(df, minLow = 5000, minTime = 10, zeropad = TRUE, ...)
}
\arguments{
\item{df}{details}

\item{minLow}{details}

\item{minTime}{details}

\item{zeropad}{details}

\item{\dots}{possibly used to create datetime variable from TimeStamp}
}
\value{
details
}
\description{
description
}
\details{
details
}
