% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markDelivery.R
\name{markDelivery}
\alias{markDelivery}
\title{Mark Delivery Days}
\usage{
markDelivery(data, cts = getOption("pa.cts"), markingString = "w",
  window = c("trim", "consecutive", "valid"), method = c("95", "mean",
  "sd"), validCut = getOption("pa.validCut"), wearThreshold = 300)
}
\arguments{
\item{data}{Data with classified wear (nonwear) status by
\code{\link{wearingMarking}}.}

\item{cts}{The name of the counts column. The default is "axis1".}

\item{markingString}{Option for summarizing wear (markingString = "w") or
nonwear time (markingString = "nw").}

\item{window}{character.  Should be one of 'trim', 'consecutive', or 'valid'.}

\item{method}{character.  Should be one of '95', 'mean' or 'sd'.}

\item{validCut}{A cutoff for the total minutes of classified monitor wear
time per day to be considered as a valid monitor day.}

\item{wearThreshold}{A numeric value specifying a pseudo-valid day cutoff
similar to "validCut".}
}
\value{
A dataframe with summary information about daily counts.
}
\description{
This function adds an indicator variable for deliver days.
}
\examples{
data(dataSec)

mydata1m = dataCollapser(dataSec, TS = "TimeStamp", col = "counts", by = 60)

data1m = wearingMarking(dataset = mydata1m,
                       perMinuteCts = 1,
                       cts = "counts")

markDelivery(data1m, cts='counts')
markDelivery(data1m, cts='counts', window='v', method='m')
markDelivery(data1m, cts='counts', window='c', method='m')
markDelivery(data1m, cts='counts', method='m')
markDelivery(data1m, cts='counts', method='s')
}

