% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking.R
\name{marking}
\alias{marking}
\title{Marking Data with Wearing Tags}
\usage{
marking(dataset, frame, cts = "axis1", streamFrame = NULL,
  allowanceFrame = 2, newcolname = "wearing")
}
\arguments{
\item{dataset}{The source dataset, in dataframe format, which needs to be
marked.}

\item{frame}{The size of time interval to be considered.}

\item{cts}{The name of the counts column. The default is "axis1".}

\item{streamFrame}{The size of time interval that the program will look back
or forward if activity is detected. The default is the half of frame.}

\item{allowanceFrame}{The size of time interval that zero counts allowed.
The default is 2.}

\item{newcolname}{The wearing marking column name.  The default is "wearing".
After the data is processed, a new field will be added to the original
dataframe.  This new field is the wearing /nowwearing indicator.}
}
\value{
A dataframe with an extra wearing/non-wearing marking column.
}
\description{
This function marks the dataset with wearing/non-wearing tags.
}
\author{
Zhouwen Liu \email{zhouwen.liu@Vanderbilt.Edu}
}
\references{
Choi, L., Liu, Z., Matthews, C. E. and Buchowski, M. S.  (2010).
Validation of Accelerometer Wear and Nonwear Time Classification Algorithm.
Med Sci Sports Exerc. Jun 23 [Epub ahead of print].
}
\keyword{internal}

