% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markingTime.R
\name{markingTime}
\alias{markingTime}
\title{Mark Days}
\usage{
markingTime(dataset, timestamp, startTime = "00:00:00",
  endTime = "23:59:59")
}
\arguments{
\item{dataset}{The source dataset, in dataframe format, that needs to be
marked.}

\item{timestamp}{The column name in the dataset that will be used as
timestamp.}

\item{startTime}{Define the starting time of a day.  It must be in the format
of "hh:mm:ss".}

\item{endTime}{Define the ending time of a day.  It must be in the format of
"hh:mm:ss".}
}
\value{
A dataframe with an extra day marking column.
}
\description{
This function adds a "day" variable to the source dataset. The day is marked
in numeric order, according to the timestamp variable.
}
\author{
Zhouwen Liu \email{zhouwen.liu@Vanderbilt.Edu}
}
\references{
Choi, L., Liu, Z., Matthews, C. E. and Buchowski, M. S.  (2010).
Validation of Accelerometer Wear and Nonwear Time Classification Algorithm.
Med Sci Sports Exerc. Jun 23 [Epub ahead of print].
}

