% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nthOccurance.R
\name{nthOccurance}
\alias{nthOccurance}
\title{Internal Function}
\usage{
nthOccurance(dataVct, value, nth = NA, reverse = FALSE)
}
\arguments{
\item{dataVct}{DataVct}

\item{value}{Value}

\item{nth}{Nth}

\item{reverse}{Reverse}
}
\value{
loc
}
\description{
This is an internal function, not for users.
}
\author{
Zhouwen Liu \email{zhouwen.liu@vumc.org}
}
\keyword{internal}
