% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicalActivity-package.R
\docType{package}
\name{PhysicalActivity-package}
\alias{PhysicalActivity}
\alias{PhysicalActivity-package}
\title{Process Physical Activity Accelerometer Data}
\description{
It contains functions to classify monitor wear and nonwear time
intervals in accelerometer data collected to assess physical activity in
free-living condition. It also contains functions to make plot for
accelerometer data, and to obtain the summary of daily monitor wear time and
the mean of monitor wear time during valid days. A monitored day is
considered valid if the total minutes of classified monitor wear time per day
is greater than a user defined cutoff.
The revised version 0.2-1 improved existing functions in terms of speed and
robustness. In addition, several functions were added such as markDelivery
that can classify days for ActiGraph delivery by mail, and markPAI that can 
categorize physical activity intensity level based on user-defined cut-points 
for accelerometer counts. It also supports importing ActiGraph agd files 
with the readActigraph and queryActigraph functions. The package also better 
supports time zones and daylight saving.
}
\details{
Classify wear and nonwear time status for accelerometer data by
epoch-by-epoch basis by \code{\link{wearingMarking}}.
}
\examples{
data(dataSec)

mydata1m = dataCollapser(dataSec, TS = "TimeStamp", col = "counts", by = 60)
options(pa.cts = 'counts') # change cnt variable from "axis1" to "counts"
data1m = wearingMarking(dataset = mydata1m, frame = 90)

sumVct(data1m, id="sdata1m")

plotData(data=data1m)

summaryData(data=data1m, validCut=600, perMinuteCts=1, markingString = "w")
}
\references{
Choi L, Liu Z, Matthews CE, Buchowski MS. 
Validation of accelerometer wear and nonwear time classification algorithm.
Med Sci Sports Exerc. 2011 Feb;43(2):357-64.

Choi L, Ward SC, Schnelle JF, Buchowski MS. 
Assessment of wear/nonwear time classification algorithms for triaxial accelerometer.
Med Sci Sports Exerc. 2012 Oct;44(10):2009-16.

Choi L, Chen KY, Acra SA, Buchowski MS. 
Distributed lag and spline modeling for predicting energy expenditure from 
accelerometry in youth. J Appl Physiol. 2010 Feb;108(2):314-27.
}
\author{
Leena Choi \email{leena.choi@Vanderbilt.Edu},
Cole Beck \email{cole.beck@vumc.org},
Zhouwen Liu \email{zhouwen.liu@vumc.org},
Charles E. Matthews \email{Charles.Matthews2@nih.gov}, and
Maciej S. Buchowski \email{maciej.buchowski@Vanderbilt.Edu}

Maintainer: Leena Choi \email{leena.choi@Vanderbilt.Edu}
}
\keyword{accelerometer}
\keyword{nonwear}
\keyword{process}
